agent/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-cartesia
python-dotenv
